Changes to the JPA annotations to get the HBM-generated schema and the annotation-generated schema to match up more closely.
git-svn-id: https://svn.code.sf.net/p/specify/code/trunk/Specify@1312 ce9bf4bf-376f-1848-a496-a1c9fe3b1a39

+++ src/edu/ku/brc/specify/datamodel/Container.java
@@ -56,7 +56,7 @@ public class Container extends DataModelObjBase implements java.io.Serializable 
     // Fields
 
      protected Long containerId;
-     protected Integer collectionObjectId;
+     protected Long collectionObjectId;
      protected Short type;
      protected String name;
      protected String description;
@@ -139,11 +139,11 @@ public class Container extends DataModelObjBase implements java.io.Serializable 
      *
      */
     @Column(name = "CollectionObjectID", unique = false, nullable = true, insertable = true, updatable = true)
-    public Integer getCollectionObjectId() {
+    public Long getCollectionObjectId() {
         return this.collectionObjectId;
     }
 
-    public void setCollectionObjectId(Integer collectionObjectId) {
+    public void setCollectionObjectId(Long collectionObjectId) {
         this.collectionObjectId = collectionObjectId;
     }
